Fix the broken approach to conditional module inclusion.
In cases where an entire class was added in a specific Qt version, we were
excluding the corresponding Hoppy module from being generated.  This isn't
sufficient because the generated modules were still being listed unconditionally
in qtah.cabal, so building for old versions would fail.

It would be nice if we could add a version range check to qtah.cabal as well,
but flags are binary and I don't think this can be done nicely.  So instead, we
now generate the module unconditionally, and just make it have no contents.

The only case in Qtah so far is QFormLayout.