Split the C++ library off into its own package.
This way, the Haskell library can link against the C++ library in a
pre-installed, known location.  The qtah-cpp package has a dummy empty
library and module so that the qtah package can get a path to the libdir
of the former (where it also installs libqtah.so.0).

This works, as long as executables using qtah are linked dynamically
(cabal configure --enable-executable-dynamic).  Otherwise, the RUNPATH
from the qtah package's libHSqtah*.so to qtah-cpp doesn't get preserved,
so the executable can't find libqtah.so.0 when run.