Add a script for creating -qtX versions of Qtah, and install script commands.
This adds a new rule to Qt version selection: if qtah-cpp or qtah's package
names end in -qtX, then Qt version X is used, regardless of QTAH_QT or package
flags.

install.sh and clean.sh now recognize -qtX versions of packages.  install.sh now
also accepts the argument "examples" to instruct it to build qtah-examples, as
well as "sdist" which makes it build a source distribution of all packages
(including the examples).