qtah-generator: Don't assume a build starts in the project root (issue #14).
qtah-generator's build process both invokes qtah-listener-gen and
installs it.  We were assuming that the initial directory for the build
was the project root and searching for qtah-listener-gen in ".", but
this seems to break with GHC 8 / Cabal 1.24 when doing "cabal install
qtah-generator" from Hackage.

This switches to calculating the project root based on the pkgDescrFile
field in LocalBuildInfo.  Hopefully this is always present and usable.