Restructure the event registration code.
The code for managing registered events was needlessly complex (e.g. it
was using signals) in order to handle unregistration or the event receiver
being deleted.  This approach also doesn't work for QGraphicsScene events,
because QGraphicsItems don't emit signals.

But, it's simpler to just have the listener be a child of the receiver.
This way, we get notification of deletion of the receiver through the
the listener's dtor, and we can still unregister (and delete) the listener
from Haskell.