Install libqtah.so to dynlibdir, and qtah-qt-version to datadir.
GHC and Cabal now have separate directories for static and dynamically
built libraries.  Installed libs are now split, and GHC seems to require
dynamic libs to be in the dynamic directory, not simply injected from the
static directory via extraLibDirs, for dependent packages to be able to use
TemplateHaskell, see [1].

This moves libqtah.so to be installed into the dynamic library directory,
and updates qtah to link against it there.  We also have to move the
specification of "extra-libraries: qtah" into qtah/Setup.hs.  There aren't
separate "extra-dynamic-libraries" or "extra-dynamic-lib-dirs" options for
Cabal files yet ([1] mentions this), so we keep using the current ones.

Finally, the qtah-qt-version file doesn't really belong in libdir, so this
moves it to datadir.

[1] https://gitlab.com/khumba/qtah/merge_requests/13