Move qtah's "extra-libraries: qtah" to Setup.hs.
This is a required part of moving to GHC and Cabal's new dynamic library
directory split[1], and also fixes an issue with cabal2nix (and possibly
other systems parsing qtah.cabal), because the setup file is needed to
actually be able to find libqtah.so.  It seems better to inject the
dependency on libqtah.so at the same time as saying where it can be
found, not before.

[1] https://gitlab.com/khumba/qtah/issues/27
[2] https://gitlab.com/khumba/qtah/merge_requests/13